Refactor BE auth | Fix Docker

+++ just4fun-be/routes/users.ts
@@ -1,53 +1,76 @@
 import express = require('express')
 import {isUser, User} from "../models/User";
 import * as user from "../models/User";
-import { express_jwt_auth } from "../bin/authentication";
+import { express_jwt_auth, passport_auth } from "../bin/authentication";
 import * as passport from "passport";
 
 let router = express.Router();
 
 //TODO restrict to admins else 403
 
-router.get('/', (req, res, next) => {
-    user.getModel().find({}, {digest:0, salt:0, isPasswordTemporary:0, avatar:0}).then( (users)=>{
+router.get('/', passport_auth(['anonymous', 'jwt']), (req, res, next) => {
+    let projection = {
+        digest:0, //Security reason
+        salt:0,   //Security reason
+        roles:0,  //Security reason
+        isPasswordTemporary:0, //Security reason
+        avatar:0  //Size of response reason
+    };
+    if(!req.user || !(req.user as User).hasAdminRole())
+    {
+        //Information available only to moderator
+        projection['following'] = 0;
+        projection['friends'] = 0;
+        projection['friendRequests'] = 0;
+        projection['roles'] = 0;
+    }
+
+    user.getModel().find({}, projection).then((users) => {
         return res.status(200).json(users);
-    }).catch( (reason)=>{
-        return next({statusCode:500, error:true, errormessage:"DB error: "+ reason});
+    }).catch((err) => {
+        return next({statusCode:500, error:true, errormessage:"DB error: "+err.errormessage});
     })
 })
 
 router.get('/leaderboard', (req, res, next) => {
     user.getModel().find({}, {digest:0, salt:0, isPasswordTemporary:0, avatar:0}).limit(10).sort({points: -1}).then( (users)=>{
         return res.status(200).json( users );
-    }).catch( (reason)=>{
-        return next({statusCode:500, error:true, errormessage:"DB error: "+ reason});
+    }).catch((err) => {
+        return next({statusCode:500, error:true, errormessage:"DB error: "+err.errormessage});
     })
 })
 
-router.get('/:email', (req, res, next) => {
-    user.getModel().findOne( {email: req.params.email}, {digest:0, isPasswordTemporary:0, salt:0} ).then( (user)=>{
+router.get('/:email', passport_auth(['anonymous', 'jwt']), (req, res, next) => {
+    let projection = {
+        digest:0, //Security reason
+        salt:0,   //Security reason
+        isPasswordTemporary:0, //Security reason (if it's your account and the pass is temp you cannot access this page so the information is not relevant)
+    };
+    if(!req.user || (req.user.email != req.params.email && !(req.user as User).hasModeratorRole()))
+    {
+        //Information only for owner or moderator
+        projection['roles'] = 0;
+        projection['following'] = 0;
+        projection['friends'] = 0;
+        projection['friendRequests'] = 0;
+        projection['roles'] = 0;
+    }
+
+    user.getModel().findOne({email: req.params.email}, projection).then((user) => {
         return res.status(200).json(user);
-    }).catch( (reason)=>{
-        return next( {statusCode:500, error: true, errormessage:"DB error"+reason} );
+    }).catch((err) => {
+        return next( {statusCode:500, error: true, errormessage:"DB error"+err.errormessage} );
     })
 });
 
-router.post('/', (req, res, next) => {
-    if (!req.body.email || req.body.email === ""){
+router.post('/', passport_auth(['anonymous', 'jwt']), (req, res, next) => {
+    if (!req.body.email || req.body.email === "")
         return next({statusCode:400, error:true, errormessage:"Mail field required"});
-    }
-    if (!req.body.name || req.body.name === ""){
-        return next({statusCode:400, error:true, errormessage:"Name field required"});
-    }
-    if (!req.body.password || req.body.password === ""){
+    if (!req.body.password || req.body.password === "")
         return next({statusCode:400, error:true, errormessage:"Password field required"});
-    }
-    if (!req.body.avatar || req.body.avatar === ""){
-        return next({statusCode:400, error:true, errormessage:"avatar field required"});
-    }
 
     let u: User;
-    if(false) //Create moderator
+    if(req.body.moderator && req.user && (req.user as User).hasModeratorRole()) //Create moderator
     {
         u = user.newUser(req.body.email, "", "");
         u.setPassword(req.body.password, true);
@@ -55,35 +78,45 @@ router.post('/', (req, res, next) => {
     }
     else
     {
+        if (!req.body.name || req.body.name === "")
+            return next({statusCode:400, error:true, errormessage:"Name field required"});
+        if (!req.body.avatar || req.body.avatar === "")
+            return next({statusCode:400, error:true, errormessage:"Avatar field required"});
+
         u = user.newUser(req.body.email, req.body.name, req.body.avatar);
         u.setPassword(req.body.password);
     }
 
-    u.save().then( (data=>{
-        return next( {statusCode: 201, error:false, errormessage:"", _id:data._id});
-    })).catch( (reason)=>{
-        if (reason.code === 11000)
-            return next( {statusCode: 400, error: true, errormessage: "User already exists"} );
-        return next( {statusCode: 500, error: true, errormessage: "DB error: "+reason.errmsg} );
+    u.save().then((data) => {
+        return next({statusCode: 201, error:false, errormessage:"", _id:data._id});
+    }).catch((err) => {
+        if (err.code === 11000)
+            return next({statusCode: 400, error: true, errormessage: "User already exists"});
+        return next({statusCode: 500, error: true, errormessage: "DB error: "+err.errormessage});
     })
 });
 
-router.delete('/:email', express_jwt_auth, (req, res, next)=>{
-    // TODO: handle authorization
-    user.getModel().deleteMany( {email: req.params.email} ).then( (user)=>{
-        return res.status(200);
-    }).catch( (reason)=>{
-        return next( {statusCode:500, error: true, errormessage:"DB error"+reason} );
-    })
+router.delete('/:email', passport_auth('jwt'), (req, res, next)=>{
+    if((req.user as User).hasModeratorRole()) {
+        user.getModel().deleteMany({email: req.params.email}).then((user) => {
+            return next({statusCode:200, error: false, errormessage:""});
+        }).catch((err)=>{
+            return next({statusCode:500, error: true, errormessage:"DB error"+err.errormessage});
+        })
+    }
+    else
+    {
+        return next({statusCode:403, error: true, errormessage:"You must be a moderator"});
+    }
 });
 
-router.put("/:id", passport.authenticate(['basic', 'jwt'], {session: false}), (req, res, next) => {
+router.put("/:id", passport_auth(['basic', 'jwt']), (req, res, next) => {
     if (req.params.id !== req.user.email)
         return next({statusCode: 403, error: true, errormessage: "Forbidden"});
 
     let acceptedFields = ['username', 'avatar', 'password'];
     let haveSetAllFields = true;
-    user.getModel().findOne({email: req.user.email}).then( (u: User) => {
+    user.getModel().findOne({email: req.user.email}).then((u: User) => {
         for (let f in acceptedFields) {
             let field = acceptedFields[f];
             if(req.body.hasOwnProperty(field))
@@ -98,40 +131,22 @@ router.put("/:id", passport.authenticate(['basic', 'jwt'], {session: false}), (r
             u.isPasswordTemporary = false;
         u.save().then(() => {
             return next({statusCode: 200, error: false, errormessage: ""});
-        }).catch(err => {
-            return next({statusCode: 500, error: true, errormessage: err});
+        }).catch((err) => {
+            return next({statusCode: 500, error: true, errormessage: "DB error: "+err.errormessage});
         });
-    }).catch(err => {
-        return next({statusCode: 500, error: true, errormessage: err});
+    }).catch((err) => {
+        return next({statusCode: 500, error: true, errormessage: "DB error: "+err.errormessage});
     });
-
-    // user.getModel().findOne({email: req.user.email}).then( async data => {
-    //     // TODO: eliminare questo in seguito che serve per debug
-    //     if (req.body.resetFriends) {
-    //         data.friends = [];
-    //         data.following = [];
-    //         data.save();
-    //         next({statusCode: 200, error: false, errormessage: ""});
-    //     }
-    //     // TODO: debug, da eliminare
-    //     else if (req.body.points) {
-    //         data.updatePoints(req.body.points, res, next)
-    //     }
-    //     else
-    //         next({statusCode: 400, error: true, errormessage: "Bad request"});
-    // }).catch(err => {
-    //     next({statusCode: 400, error: true, errormessage: err});
-    // });
 });
 
 router.post('/:id/follow', express_jwt_auth, (req, res, next) => {
     if (req.params.id !== req.user.email)
         return next({statusCode: 403, error: true, errormessage: "Forbidden"});
 
-    user.getModel().findOne({email: req.user.email}).then(data => {
+    user.getModel().findOne({email: req.user.email}).then((data) => {
         data.follow(req.body.user, res, next);
-    }).catch(err => {
-        next({statusCode: 400, error: true, errormessage: err});
+    }).catch((err) => {
+        return next({statusCode: 400, error: true, errormessage: "DB error: "+err.errormessage});
     });
 });
 
@@ -139,10 +154,10 @@ router.delete('/:id/follow', express_jwt_auth, (req, res, next)=>{
     if (req.params.id !== req.user.email)
         return next({statusCode: 403, error: true, errormessage: "Forbidden"});
 
-    user.getModel().findOne({email: req.user.email}).then(data => {
+    user.getModel().findOne({email: req.user.email}).then((data) => {
         data.unfollow(req.body.user, res, next);
-    }).catch(err => {
-        next({statusCode: 400, error: true, errormessage: err});
+    }).catch((err) => {
+        return next({statusCode: 500, error: true, errormessage: "DB error: "+err.errormessage});
     });
 });
 
@@ -150,12 +165,10 @@ router.post('/:id/friend', express_jwt_auth, (req, res, next) => {
     if (req.params.id !== req.user.email)
         return next({statusCode: 403, error: true, errormessage: "Forbidden"});
 
-    user.getModel().findOne({email: req.user.email}).then(data => {
-        let u: User;
-        if (isUser(data)) u = data;
+    user.getModel().findOne({email: req.user.email}).then((data) => {
         data.sendFriendRequest(req.body.user, res, next);
-    }).catch(err => {
-        next({statusCode: 400, error: true, errormessage: err.message});
+    }).catch((err) => {
+        return next({statusCode: 500, error: true, errormessage: "DB error: "+err.errormessage});
     });
 })
 
@@ -163,13 +176,15 @@ router.put('/:id/friend', express_jwt_auth, (req, res, next) => {
     if (req.params.id !== req.user.email)
         return next({statusCode: 403, error: true, errormessage: "Forbidden"});
 
-    user.getModel().findOne({email: req.user.email}).then(data => {
+    user.getModel().findOne({email: req.user.email}).then((data) => {
         if (req.body.accept)
             data.acceptFriendRequest(req.body.accept, res, next);
         else if (req.body.refuse)
             data.refuseFriendRequest(req.body.refuse, res, next)
-    }).catch(err => {
-        next({statusCode: 400, error: true, errormessage: err});
+        else
+            return next({statusCode: 400, error: true, errormessage: "Specify accept or refuse"});
+    }).catch((err) => {
+        return next({statusCode: 500, error: true, errormessage: "DB error: "+err.errormessage});
     });
 })
 
@@ -177,12 +192,11 @@ router.delete('/:id/friend', express_jwt_auth, (req, res, next) => {
     if (req.params.id !== req.user.email)
         return next({statusCode: 403, error: true, errormessage: "Forbidden"});
 
-    user.getModel().findOne({email: req.user.email}).then(data => {
+    user.getModel().findOne({email: req.user.email}).then((data) => {
         data.removeFriend(req.body.friend, res, next);
-    }).catch(err => {
-        next({statusCode: 400, error: true, errormessage: err});
+    }).catch((err) => {
+        return next({statusCode: 500, error: true, errormessage: "DB error: "+err.errormessage});
     });
-
 })
 
 module.exports = router;
